Add finer-grained control of Rake tasks loading/running
The old approach was all or nothing. So, if the project didn't want
`yard` tasks defined and run, then they couldn't opt out of this very
easily. The goal here is to allow for full control and definition of
which tasks to load and run within the including project, itself.