Fix eslint.rake to actually find/output/fail on errors
Best I can tell: within the rake task's environment, the `**/*.js` wild
card is not "properly" resolving. It's at least not resolving in the
way that eslint is expecting... or something. But I noticed that running
the same command from the shell produced errors while it didn't from
rake tasks (in another project). Then I just poked it with a stick until
I found something that worked in both environments.

BEFORE FIX: Example output of a failure:

```
> rake eslint
0 errors, 0 warnings
```

AFTER FIX: Example output of a failure:

```
> rake eslint

/Users/paul/dev/minesweeper_alliance/app/javascript/application.js
   5:1   error  'Turbo' is not defined  no-undef
  11:40  error  'Turbo' is not defined  no-undef

✖ 2 problems (2 errors, 0 warnings)
```