Ignore rubocop/reek warnings in `create`/`update` actions
I'm convinced that as soon as one needs to add a
`respond_to do |format|` block to a controller action they're then going
to hit Metrics limits soon thereafter. Some attention to this is good,
yes.... But the pressure is driving me to want to do abstractions that
I shouldn't actually be doing.

So it seems better to just turn off these metrics (for these action
methods and only in the Rails config). We'll just call them exceptions
to the rule for standard Ruby practice.