Fix `brakeman` task reporting on ignored warnings
Previously, the `brakeman` task was reporting on ignored warnings
(a.k.a. filtered warnings).

This made it feel like an issue was present when it wasn't, since it
caused a wall of text compared to the usual `0 warnings`.

Also, don't try to output "Warnings found" when there are warnings to
report on, as it just gets lost at the top of the large wall of text
(the report).