Enable all cops by default + Update config for new additions
I didn't realize before that we weren't getting new cops by default
(cops that are new are marked as Enabled: pending, and we don't get
these by default). This can be disabled on a per-project basis, if
desired, but this is the only way I can learn about new additions.