Tweaks to rake tasks: style + light refactoring
The largest change herein is that we no longer "take arguments" on the
brakeman task. Though I don't know how it was suppose to work before.
Basically was just kind of useless so removing it.

Also, change `exit 1` to just `abort` which does the same thing but is
a bit more clear on what it's doing. (e.g. is `1` good or bad?)