Improve output on error in `rake prettier` task
Previously, when the prettier failed, it wasn't very clear that it
wasn't just some sort of gem error or something... because a stack
trace-looking thing was being output.

To fix: we use `system` and intentionally return `abort` when the
check isn't successful--as we do in other rake tasks. This fixes the
symptom, freeing the user up to read the error message from the prettier
check's output.