Revert: Rubocop: ignore all files in the .gitignore spec
When used from a project, this line was resulting in an git error on
every rubocop invocation:
`fatal: not a git repository (or any of the parent directories): .git`

So I guess if we're going to use this it'll have to be on a per-project
basis, where the actual git files are. (I'm assuming it's because there
is no git project within an installed gem. i.e. it's not sourcing the
project, but the gem root.)

Note: I didn't add the exclusions for `not_modules` and `tmp` in the
non-Rails version of the rubocop/all_cops.yml because I don't expect
my gems to have these paths.