Disable Metrics/BlockLength cop in Tests
Previously, Rubocop's Metrics/BlockLength was cop complaining about a
`given` block being too long in an implementing gem.

To fix: Instead of just adding another method name to ignore to the list
of "AllowedMethods", it makes sense to just ignore test files
altogether, as we do for the Rails version of this cop.

Also, fix alphabetization of cops in lib/rubocop/metrics-rails.yml